Render points with per-fragment depth.
Each point in a point cloud is now rendered with depth that is
approximately a full sphere, rather than a flat camera-facing billboard.
This is a very approximate calculation, but one which is good as long as
the sphere is not too off-axis; it's the same kind of approximation as
already used by the brightness `shading` of the points. The primary goal
of this change is to make “ball and stick” constructions of point and
line entities look better.

I intended this algorithm to be equally applicable to lines (making them
render as cylinders), but that is not implemented yet.