Ensure we have roundtrip tests for each archetype in PR CI (#6524)
`main` CI is failing:
https://github.com/rerun-io/rerun/actions/runs/9436507239/job/25991550784

https://github.com/rerun-io/rerun/pull/6510 added a new archetype
(`Axes3D`) without adding roundtrip tests for it. This was easily missed
because the PR CI doesn't run the roundtrip tests.

This PR adds a fast PR CI check to make sure we have roundtrip tests for
all archetype, but without paying the cost of actually running them. The
failure looks like this:


![image](https://github.com/rerun-io/rerun/assets/1148717/61bf60b6-4a88-4f13-be65-917741d7aff8)

This PR also adds an exception for the `Axes3D` archetype. I've created
a separate issue for that, adding it to this cycle:
* https://github.com/rerun-io/rerun/issues/6525

---

* [x] I love checkboxes