Remove `panic="abort"` from `dev` builds
The `backtrace` crate isn't compatible with `panic="abort"` (see
[docs](https://docs.rs/backtrace/latest/backtrace/) and
[issue](https://github.com/rust-lang/backtrace-rs/issues/397)).

One great feature we have in egui is pressing down all modifier keys
on your keyboard and seeing the backtrace to the widget you hover.
This is awesome for when you want to change the code for some
specific part of the UI. But it didn't work, at least not always.

This PR fixes it.

We still use `panic="abort"` in release builds, for the benefit
of smaller binaries.

* See also https://github.com/emilk/egui/pull/4696