New temporal batch APIs (#6587)
- Resolves: https://github.com/rerun-io/rerun/issues/4184

This primarily introduces a new logging API for temporal batches. This
API is a slightly lower-level than the existing `log` API but has a
fairly familiar feel to it if you've worked with our data-types.

The biggest difference is that it does not currently support Archetypes.
Data must be specifically logged using raw components arranged in
(possibly partitioned) batches. The main reason for this is that
Archetypes aren't required to have matched-length components and as such
you would need to provide a per-component parititioning, which starts to
look very similar to the manual component-level API.

Note that we automatically wrap regular batches in the correct way to
turn them into "mono-batches". Partitions only need to be generated
manually to support batch-batches.

This also requires a few helper classes for the 3 `TimeBatch` types, but
otherwise makes use of the existing `ComponentBatch` constructors.


### New python API docstring:
```python
def log_temporal_batch(
    entity_path: str,
    times: Iterable[TimeBatchLike],
    components: Iterable[ComponentBatchLike],
    recording: RecordingStream | None = None,
    strict: bool | None = None,
) -> None:
```
  Directly log a temporal batch of data to Rerun.

Unlike the regular `log` API, which is row-oriented, this API lets you
submit the data
in a columnar form. Each `TimeBatchLike` and `ComponentBatchLike` object
represents a column
of data that will be sent to Rerun. The lengths of all of these columns
must match, and all
data that shares the same index across the different columns will act as
a single logical row,
  equivalent to a single call to `rr.log()`.

Note that this API ignores any stateful time set on the log stream via
the `rerun.set_time_*` APIs.

When using a regular `ComponentBatch` input, the batch data will map to
single-valued component
  instance at each timepoint.

  For example, scalars would be logged as:
  ```py
  times = np.arange(0, 64)
  scalars = np.sin(times / 10.0)

  rr.log_temporal_batch(
      "scalars",
      times=[rr.TimeSequenceBatch("step", times)],
      components=[rr.components.ScalarBatch(scalars)],
  )
  ```
In the viewer this will show up as 64 individual scalar values, one for
each timepoint.

However, it is still possible to log temporal batches of batch data. To
do this the source data first must
be created as a single contiguous batch, and can then be partitioned
using the `.partition()` helper on the
  `ComponentBatch` objects.

For example, to log 5 batches of 20 point clouds, first create a batch
of 100 (20 * 5) point clouds and then
  partition it into 5 batches of 20 point clouds:
  ```py
  times = np.arange(0, 5)
  positions = rng.uniform(-5, 5, size=[100, 3])

  rr.log_temporal_batch(
      "points",
      times=[rr.TimeSequenceBatch("step", times)],
      components=[
          rr.Points3D.indicator(),
rr.components.Position3DBatch(positions).partition([20, 20, 20, 20,
20]),
      ],
  )
  ```
In the viewer this will show up as 5 individual point clouds, one for
each timepoint.


---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>