Compaction-aware store events (#6940)
This makes store events aware of compaction, i.e. when a new chunk comes
in and gets compacted with an older chunk, downstream subscribers will
now be notified of what the new chunk looked like _pre-compaction_, and
which chunk was merged with what.

This results in massive perf boost in downstream subscribers without
requiring code changes (existing subscribers will effectively be working
with diffs without knowing).