New chunkified latest-at APIs and caches (#6992)
_(Don't be fooled by the diff size, it is 90% copy pasta)_

Introduces the new chunkified latest-at APIs and related caches.

This is implemented, and works, very similarly to the legacy one.
The one major difference is that we store `UnitChunk`s, and never ever
deserialize data! :partying_face:

In addition, the new cache maintain a simple "`Clear` tracker", which
greatly improved the performance of read-time clears.
I've backported this mechanism to the legacy re_query, because why not.