Combine `re_types` integration tests into one test target.
This significantly reduces (re)build times and total artifact size
for these tests as a whole.

A disadvantage is that now `tests/types/main.rs` needs to list child
modules. If desired, this could be added to the code generator; the
`cfg`s in `main.rs` could equally well be moved inside each module
to make the file completely regular.