Add `ProcMeshVisBuffer`, deduplicating visualizer code.
`ProcMeshVisBuffer` extracts common elements from `Boxes3DVisualizer`
and `Ellipsoids3DVisualizer`.

Additionally, `ComponentFallbackProvider` is no longer a supertrait of
`VisualizerSystem`. Instead, the fallback provider is always accessed
through `as_fallback_provider()` (which may return `&self` if it likes).
The advantage of this is that a visualizer may access its desired
fallback logic even while parts of itself are exclusively borrowed, by
putting the `TypedComponentFallbackProvider` implementations on a
separate type. (Strictly speaking, the supertrait change is not needed,
if we instead let a `ComponentFallbackProvider` implementation be made
from `TypedComponentFallbackProvider`s on a different type, but that
seems like a worse design to me.)