Retry reading .rrd files until end-of-stream message header (#7475)
Ensure we continue streaming .rrd file in the viewer as it's being written to (#4056)

We introduce a retryable file reader that is used when reading .rrd
files that are still being written and it simply ignored the EoF
condition. We also introduce an end-of-stream message header that
ensures we stop decoding (and endlessly trying to read the file) message
stream once we see this header.

.rrd files generated with using a prior sdk version should work just
fine, the only downside is that while the viewer is running there will
be an open file handle for that file since no end-of-stream message will
exist.