Chunk store: support for overlapped range queries (#7586)
Turns out I never got around to implementing support for overlapped
chunks in range queries in the `ChunkStore`.

I only realized after hunting down a crazy bug in the dataframe APIs for
way too long...

The fix is the exact same as the one used for latest-at queries. In fact
it is a copy paste.