Show list of enabled feature with `rerun --version` (#7885)
* Reverts rerun-io/rerun#7770
* Re-applies https://github.com/rerun-io/rerun/pull/7744

The revert was because during `cargo publish`, a `Cargo.lock` file
materialized, and that made cargo angry:

```
Failed to publish re_sdk:
error: failed to verify package tarball

Caused by:
  Source directory was modified by build.rs during cargo publish. Build scripts should not modify anything outside of OUT_DIR.
  Added: /home/runner/work/rerun/rerun/target/package/re_sdk-0.19.0-alpha.9/Cargo.lock
```

So in this PR we only query about which cargo features if we're NOT in a
`cargo publish` build. That should work! We'll know when we next publish
some crates :)