Remove the upper bound constraint on python version
We previously had this constraint in place to avoid certain edge cases
of users not being able to install Rerun and not understanding why due
to transitive deps. Primarily on pyarrow.

I've confirmed pyarrow is now doing something similar in their own
project:
https://github.com/apache/arrow/blob/main/python/pyproject.toml#L37

Let's try to let the package solver do the right thing and only add a
max-version if we find we are truly incompatible with some future python
version.