Port codegened arrow serialization to arrow1 (#8208)
* Follows https://github.com/rerun-io/rerun/pull/8206
* Part of https://github.com/rerun-io/rerun/issues/3741

## Changes
To implement nullable unions, we have a `_null_marker: Null` variants in
all our unions. This means all our unions are nullable.

Previously we would only mark a struct field as nullable if it was
declared as such in the `.fbs` file, but `arrow-rs` complains about
this. So with this PR, if a struct field refers to a union type, that
struct field will be marked as `nullable: true` in the datatype (in
Rust, Python and C++).