Space view screenshotting in native viewer (#8258)
### Related
* Closes https://github.com/rerun-io/rerun/issues/3941
* Uses https://github.com/emilk/egui/pull/5416
* Uses https://github.com/emilk/egui/pull/5426
* Follow-up: https://github.com/rerun-io/rerun/issues/8264

### What
Right-click the name of any space view to get the option to copy its
contents as a screenshot, or save it to disk:


![image](https://github.com/user-attachments/assets/0a0c5215-9042-4989-9e4a-62f76823588a)

-> 


![image](https://github.com/user-attachments/assets/ec222323-4692-4fd2-ae7b-2342996829a3)


#### Details
The implementation simply takes a full-screen screenshot and crops the
result.

The previous attempt (behind a feature-flag) only captured things that
were rendered with `re_renderer`, i.e. did not include any egui
elements, like labels. So it also only worked on spatial views. This PR
work on _all_ space views, and could also be extended to work on
containers etc.

### Limitations
Not yet implemented on web. See
https://github.com/rerun-io/rerun/issues/8264

### TODO
* [x] Point commit hash to egui `master`

---------

Co-authored-by: Andreas Reich <andreas@rerun.io>