Remove selection history
We're removing it because we're worried users will confuse it with
the new Undo feature.

After this PR, you can no longer go back in the selection history.

But, we're planning on adding hierarchy breadcrumbs to the selections
instead, which should hopefully be an even better way of understanding
and navigating selections:
* https://github.com/rerun-io/rerun/issues/4491