Fix various doc links in SDKs (#8331)
PR title is an accompanying changelog entry, but the real thing here is
all about codegen fixes:

* use `reporter` for doc issues
* make `reporter` output colored
* support field & enum links - this is a bit limited depending on the
language. Experimented a bit, but deemed this now good enough. We don't
use those often anyways
* fix links
* disable warnings for python autogen serialization being available when
a custom one is present - in all cases we have the custom serialization
there for a good reason or aren't entirely sure if it would break
something