Enforce blueprint data is always untagged during the interim (#8409)
We're in a transition period during which the Rerun ecosystem is slowly
moving over to tagged data.

During that time, it is common to end up in situations where the
blueprint intermixes both tagged
and untagged components, which invariably leads to undefined behavior.
To prevent that, we just always hot-patch it to untagged, for now.

Examples:
* An SDK logs a blueprint (tagged), which is then updated by the viewer
(which uses untagged log calls).
* Somebody loads an old .rbl from somewhere and starts logging new
blueprint data to it.
* Etc.

This fixes all known blueprint bugs from the last couple days.

In the future, we will want to make sure that the viewer logs blueprint
data with appropriate tags, which is part of the
`log(MyArchetype.my_component())` story.
The sooner the better, so that the ecosystem can start to converge on
always-tagged blueprint files.