Allow drag-and-dropping multiple containers and views in the blueprint tree (#8334)
### Related

* Closes #8276
* Closes #8275
* Part of #8266
* Part of #8267
* Related to #7108

### What

This PR makes it possible to drag multi-selection of views and
containers within the blueprint tree. It lays the foundation of a system
that will be extended to other drag payload and UI sections.

Specifically:
- `ctx.handle_select_hover_drag_interactions()` (formerly
`ctx.select_hovered_on_click()`) is now able to initiate drag
interactions. This is opt-in for now, as dragging from most place isn't
supported yet (to implemented in follow-up PRs).
- Introduce a new `DragAndDropPayload` type to interoperate between
various part of the UI. This type also enforce the grouping of items
that can meaningfully be dragged together (e.g. it's ok to drag a view
and a container together, because there exist somewhere they can be
dropped to, but it's not ok to drag a view and an entity together).
- When a drag is successfully initiated, a "pill" is displayed along the
cursor which indicates the content of what's being dragged.
- Introduces a _very_ hack mechanism for a black list of undraggable
items (aka the root container).
- Update blueprint tree to support multiple selection and the new drag
and drop payload type.
- Updates egui to latest `master`.

---------

Co-authored-by: Emil Ernerfeldt <emil.ernerfeldt@gmail.com>