Add missing click and hover interactions for views (#8495)
### Related

<!--
Include links to any related issues/PRs in a bulleted list, for example:
* Closes #1234
* Part of #1337
-->

* Closes #8479
* Closes #5138

### What

This adds missing view interactions the following views:

* Map view
* Graph view
* TextDocument view

@Wumpf @abey79 I feel like the pattern repeats itself and it's easy to
miss when writing new views. So I wonder if this could be unified. I see
several options for this:

* repurpose `handle_select_hover_drag_interactions` which only seems to
be used for `DataResult` right now
* Making each view return a `Result<(Response,
ViewSystemExecutionError>` (instead of `()`) and handling it outside of
the view.
* Creating a function analogous to
`handle_select_hover_drag_interactions`.

<!--
Make sure the PR title and labels are set to maximize their usefulness
for the CHANGELOG,
and our `git log`.

If you have noticed any breaking changes, include them in the migration
guide.

We track various metrics at <https://build.rerun.io>.

For maintainers:
* To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.
* To deploy documentation changes immediately after merging this PR, add
the `deploy docs` label.
-->