Add `MainThreadToken` to ensure file-dialogs only run on the main thread (#8467)
### What

We have a foot-gun in our code: our file dialogs (via `rfd`) [are only
allowed to be run from the main thread (at least on
Mac)](https://docs.rs/rfd/latest/rfd/#macos-non-windowed-applications-async-and-threading).
However, there is nothing stopping a user from accidentally calling
these functions from a background thread, and if you test it on e.g.
Linux, it may very well work.

So this PR introduces a new crate `re_capabilities` and a new type
`MainThreadToken`. Any function that uses `rfd` should require the
`MainThreadToken` as an argument. The `MainThreadToken` is only allowed
to be created in `fn main`, and since it is neither `Send` nor `Sync`,
this guarantees at compile-time that any functions that take a
`MainThreadToken` can only be called from the main thread.

NOTE: I have no way to enforce that all uses of `rfd` also require
`MainThreadToken` - we need to remember this ourselves, but I've made
sure that all our _current_ uses of `rfd` require a `MainThreadToken`.