Fix setting fallback `Position2D` and geometry improvements (#8505)
### Related

* Closes #8503.

### What

This fixes setting the fallback value in the selection panel.

It also deals with geometry that can't be drawn in a meaningful way
because of that.

<!--
Make sure the PR title and labels are set to maximize their usefulness
for the CHANGELOG,
and our `git log`.

If you have noticed any breaking changes, include them in the migration
guide.

We track various metrics at <https://build.rerun.io>.

For maintainers:
* To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.
* To deploy documentation changes immediately after merging this PR, add
the `deploy docs` label.
-->