Fix undo not working in some entity drag-and-drop instances (#8526)
This is a band-aid patch to address an issue with entity path filter
which would result in the following behaviours:

- the entity path filter would be perma-writen to the blueprint store
when the view is selected (in some cases)
- in such cases, this would break undo on user change (aka with entity
drag-and-drop)

The core issue is deeper and might cause panics with rust 1.81+. The
TL;DR is that the current `EntityPathFilter` type is ambiguous as to
whether its content has substitutions (aka $origin) applied or not. In
particular `Eq` and `Ord` apply to unsubstituted, resp. substituted
content (which can lead to the above panic). This should be further
cleaned by having two structures, one unsubstituted and another
substituted.

---------

Co-authored-by: Clement Rey <cr.rey.clement@gmail.com>