Fix tooltips being dragged along in graph view (#8573)
### Related

<!--
Include links to any related issues/PRs in a bulleted list, for example:
* Closes #1234
* Part of #1337
-->

* Closes #8568.

### What

This fixes the problem by tying to tooltip position to the node position
instead of the the pointer, with the nice side-effect that tooltips can
now be interacted with in the graph view.

There is some weirdness with tooltips opening for the tooltips, but we
should probably tackle that separately.

<img width="607" alt="image"
src="https://github.com/user-attachments/assets/43e0edba-49b6-4f49-a279-0982320c4ef6"
/>

<!--
Make sure the PR title and labels are set to maximize their usefulness
for the CHANGELOG,
and our `git log`.

If you have noticed any breaking changes, include them in the migration
guide.

We track various metrics at <https://build.rerun.io>.

For maintainers:
* To run all checks from `main`, comment on the PR with `@rerun-bot
full-check`.
* To deploy documentation changes immediately after merging this PR, add
the `deploy docs` label.
-->