Reallow `#[cfg(TODO)]` (#8585)
I very often use `#[cfg(TODO)]` as a way to semantically comment entire
scopes of code at once, safely (i.e. the CI will fail if the code is
committed as is).

Rust 1.81 broke that because `cfg(TODO)` is not referenced in the lint
database, thus failing to compile this kind of code.
This PR tells `cargo` to relax.