Autogenerate tagging-compliant descriptor methods for all archetypes (#8643)
Generate methods for all archetypes that allow retrieving the
fully-qualified descriptor for any of the archetype's field.
E.g. this method returns the descriptor for the `radii` field of
`Points3D`:
```rust
impl Points3D {
    /// Returns the [`ComponentDescriptor`] for [`Self::radii`].
    #[inline]
    pub fn descriptor_radii() -> ComponentDescriptor {
        ComponentDescriptor {
            archetype_name: Some("rerun.archetypes.Points3D".into()),
            component_name: "rerun.components.Radius".into(),
            archetype_field_name: Some("radii".into()),
        }
    }
}
```

This becomes a must as we start require tags in more and more places
(e.g. partial updates APIs).

* DNM: requires #8642 
* Part of https://github.com/rerun-io/rerun/issues/8581