Fix clicking of links in markdown (#8794)
### Related
* Closes https://github.com/rerun-io/rerun/issues/8574
* Bug introduced in https://github.com/rerun-io/rerun/pull/8495

### What
The text document had a `response.interact(Sense::click())` call _on
top_ of it, catching all clicks that should have ended up on the
hyperlinks. This is now changed into setting `Sense::click` on the
underlying `Ui` when creating it, meaning it will catch clicks _behind_
the hyperlinks, i.e. any click that doesn't interact with the document
iteself.