Include a python API for routing time control commands to the notebook instance (#8809)
### Related
- https://github.com/rerun-io/rerun/pull/8673

### What

The linked PR added support for using javascript to update the time
controls. Routing this into python when working in the context of a
notebook unlocks lots of interesting interactive behavior.

This does not yet expose the reverse of generating python callback
events based on rerun time cursor changes, but that would be an obvious
next step.