The empty/full entity icon now reflects presence of component on the current timeline (#8839)
### What

According to the code, the entity icon was supposed to reflect the
existence of data _on the current timeline_. However, due to a bug in
`all_components_on_timline`, it reflected the existence of data _on any
timeline_.

This PR fixes the aforementioned bug. As a result, the icon now
_actually_ represents the state on current timeline. Some snapshot tests
had to be fixed to pass.

**Reminder**: As soon as a static component is logged to an entity, it
is deemed to have data on _all_ past, present, and future timelines. As
a result, it always has the "full" icon.

---------

Co-authored-by: jprochazk <pjanv42@gmail.com>